backend/requirements.txt
# Process Failure Predictor Backend

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
xgboost>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
